GameLobby: Display a list of users currently connected to a room.
Implemented websocket endpoints on the client for displaying a list of
users in the gamelobby.

Can be found in the Room component.

NOTE: when run in the browser, it will try to connect to websockets on
localhost:8080. I did not find a good way to connect it to i.e.
swaggerhub.